+++ crates/wdl-engine/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Fixed
 
+* Use `IndexMap` for stable serialization of `Config` ([#602](https://github.com/stjude-rust-labs/wdl/pull/602)).
 * Fixed deserialization of `Object` to no longer require keys be WDL
   identifiers ([#586](https://github.com/stjude-rust-labs/wdl/pull/586)).
 * Fixed a panic caused by an incorrect type calculation of non-empty array